add predict

+++ tf/rnn/char-rnn/train.sh
@@ -10,8 +10,10 @@ echo 'TF_CONFIG = ' ${TF_CONFIG}
 model_dir=`pwd`/model_dir
 export_model_dir=`pwd`/export_model_dir
 
+run_mode='train'
+
 remove_model_dir=1
-if [[ ${remove_model_dir} == '1' ]]; then
+if [[ ${remove_model_dir} == '1' && ${run_mode} != 'predict' ]]; then
     echo "remove ${model_dir}.bak"
     rm -rf ${model_dir}.bak
     if [[ -d ${model_dir} ]]; then
@@ -29,10 +31,10 @@ params=(\
 [preprocessed_filename]=${model_dir}/preprocessed.pkl \
  \
 ## run_mode: train, predict, all \
-[run_mode]='train' \
+[run_mode]=${run_mode} \
 [train_data_path]=`pwd`/data/tinyshakespeare/input.txt \
 [eval_data_path]=`pwd`/data/tinyshakespeare/input.txt \
-[predict_data_path]=`pwd`/validation.txt \
+[predict_data_path]=`pwd`/predict.txt \
 [predict_output]=`pwd`/predict_output.${model_name}.txt \
 [predict_checkpoint_path]=${model_dir} \
  \
@@ -45,7 +47,7 @@ params=(\
 [clip_norm]=5.0 \
 [batch_size]=32 \
 [max_train_steps]=-1 \
-[epoch]=30 \
+[epoch]=200 \
 [throttle_secs]=600 \
 [use_embedding]=False \
 [embedding_dim]=100 \
@@ -62,11 +64,11 @@ params=(\
 [save_checkpoints_steps]=-1 \
 [save_checkpoints_epoches]=1.0 \
 [keep_checkpoint_max]=5 \
-[log_step_count_steps]=10 \
+[log_step_count_steps]=100 \
  \
 # profile flags \
 [use_profile_hook]=False \
-[profile_steps]=10 \
+[profile_steps]=100 \
  \
 # optimizer flags \
 [optimizer]='adam' \
@@ -74,7 +76,7 @@ params=(\
 [adagrad_initial_accumulator_value]=0.1 \
 [adam_beta1]=0.9 \
 [adam_beta2]=0.999 \
-[opt_epsilon]=1.0 \
+[opt_epsilon]=1e-8 \
 [ftrl_learning_rate_power]=-0.5 \
 [ftrl_initial_accumulator_value]=0.1 \
 [ftrl_l1]=0.0 \
@@ -84,9 +86,9 @@ params=(\
 [rmsprop_decay]=0.9 \
  \
 # learning rate flags \
-[learning_rate]=0.01 \
+[learning_rate]=0.001 \
 ## fixed, exponential or polynomial
-[learning_rate_decay_type]='exponential' \
+[learning_rate_decay_type]='fixed' \
 [end_learning_rate]=0.0001 \
 [learning_rate_decay_factor]=0.95 \
 [num_epochs_per_decay]=2.0 \
